Add graphics and music
