+++ Pong2/MenuVC/MenuVC.swift
@@ -1,19 +1,25 @@
 import UIKit
+import AVFoundation
 
 class MenuVC : UIViewController {
+
+    let sound = AVPlayer(url: Bundle.main.url(forResource: "backgroundSound", withExtension: "mp3")!)
 
     @IBOutlet weak var online: UIButton!
     @IBOutlet weak var status: UILabel!
     //MARK: - @IBActions
     @IBAction func online(_ sender: UIButton) {
+        sound.pause()
         moveToGameVC(game: .online)
     }
 
     @IBAction func player2(_ sender: UIButton) {
+        sound.pause()
         moveToGameVC(game: .player2)
     }
 
     @IBAction func offline(_ sender: UIButton) {
+        sound.pause()
         moveToGameVC(game: .offline)
     }
     
@@ -25,6 +31,7 @@ class MenuVC : UIViewController {
     }
     
     override func viewDidLoad() {
+        sound.play()
         if  Internet.connection() == false {
             online.isEnabled = false
             status.text = "No internet connection"
